Fix MIT license for pkg.go.dev; tidy go.mod and create go.sum (#62)
Currently the documentation for github.com/kr/pretty is not displayed on
pkg.go.dev because of license restrictions:

    https://pkg.go.dev/github.com/kr/pretty?tab=doc

This is almost certainly because the License file includes an errant
header:

    The MIT License (MIT)

The actual MIT license does not include this header:

    https://opensource.org/licenses/MIT

GitHub and other interfaces surface the type of the license through
automatic detection (as will pkg.go.dev). Hence this header is
superfluous in any case.

Whilst we are at it, tidy up go.mod and add go.sum.